task-service/requirements.txt
redis
grpcio
grpcio-tools
grpcio-reflection
